refactor: marketing page structure
