+++ apps/web/src/app/(content)/play/checker/[id]/page.tsx
