+++ apps/web/src/app/(content)/play/checker/_components/checker-form.tsx
@@ -53,12 +53,11 @@ import { useQueryStates } from "nuqs";
 import { searchParamsParsers } from "../search-params";
 
 const FloatingActionNoSSR = dynamic(
-  () =>
-    import("../_components/floating-action").then((mod) => mod.FloatingAction),
+  () => import("./floating-action").then((mod) => mod.FloatingAction),
   {
     ssr: false,
     loading: () => <></>,
-  },
+  }
 );
 
 /**
@@ -177,7 +176,7 @@ export function CheckerForm({ defaultValues, defaultData }: CheckerFormProps) {
                       `Checking ${regionFormatter(_result[0].region, "long")} (${latencyFormatter(_result[0].latency)})`,
                       {
                         id: toastId,
-                      },
+                      }
                     );
                   }
                 }
