+++ apps/web/src/app/(content)/play/checker/_components/global-monitoring.tsx
