+++ apps/web/src/app/(content)/play/checker/page.tsx
