+++ apps/web/src/app/(content)/play/curl/layout.tsx
