+++ apps/web/src/app/(content)/play/curl/page.tsx
