+++ apps/web/src/app/(content)/play/default.tsx
