+++ apps/web/src/app/(content)/play/status/_components/status-play.tsx
