+++ apps/web/src/app/(content)/play/status/layout.tsx
