--- apps/web/src/app/play/layout.tsx
@@ -1,22 +1,0 @@
-import {
-  defaultMetadata,
-  ogMetadata,
-  twitterMetadata,
-} from "@/app/shared-metadata";
-import { MarketingLayout } from "@/components/layout/marketing-layout";
-import type { Metadata } from "next";
-import type { ReactNode } from "react";
-
-export const metadata: Metadata = {
-  ...defaultMetadata,
-  twitter: {
-    ...twitterMetadata,
-  },
-  openGraph: {
-    ...ogMetadata,
-  },
-};
-
-export default function PlayLayout({ children }: { children: ReactNode }) {
-  return <MarketingLayout>{children}</MarketingLayout>;
-}
